Event refactor and user limits

+++ back/src/main/java/com/ionnote/services/EventService.java
@@ -4,6 +4,8 @@ import com.ionnote.dtos.event.*;
 import com.ionnote.dtos.event.EventDTO;
 import com.ionnote.entities.Event;
 import com.ionnote.exceptions.EventNotFoundException;
+import com.ionnote.exceptions.ForbiddenException;
+import com.ionnote.exceptions.UserNotFoundException;
 import com.ionnote.repositories.EventRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -16,11 +18,14 @@ import java.util.UUID;
 @AllArgsConstructor
 public class EventService {
     private EventRepository eventRepository;
+    private AuthService authService;
 
     //Create
-    public void createEvent(CreateEventDTO dto) {
+    public void createEvent(CreateEventDTO dto) throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
         var tempEvent = Event.builder()
-                .uuid(UUID.randomUUID())
+                .id(UUID.randomUUID().toString())
+                .ownerId(loggedUser.getId())
                 .name(dto.getName())
                 .start(dto.getStart())
                 .end(dto.getEnd())
@@ -30,16 +35,21 @@ public class EventService {
     }
 
     //Read
-    public EventDTO readEvent(ReadEventDTO dto) throws EventNotFoundException {
+    public EventDTO readEvent(ReadEventDTO dto) throws EventNotFoundException, UserNotFoundException, ForbiddenException {
         var event = eventRepository.findById(dto.getUuid()).orElseThrow(EventNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(event.getId())){
+            throw new ForbiddenException();
+        }
         var response = new EventDTO();
         BeanUtils.copyProperties(event,response);
         return response;
     }
 
-    public List<EventDTO> readAllEvents() {
-        var tasks = eventRepository.findAll();
-        return tasks.stream().map(event -> {
+    public List<EventDTO> readAllEvents() throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
+        var events = eventRepository.findAllByOwnerId(loggedUser.getId());
+        return events.stream().map(event -> {
             var tempResponse = new EventDTO();
             BeanUtils.copyProperties(event,tempResponse);
             return tempResponse;
@@ -47,15 +57,23 @@ public class EventService {
     }
 
     //Update
-    public void updateEvent(UpdateEventDTO dto) throws EventNotFoundException {
+    public void updateEvent(UpdateEventDTO dto) throws EventNotFoundException, UserNotFoundException, ForbiddenException {
         var event = eventRepository.findById(dto.getUuid()).orElseThrow(EventNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(event.getId())){
+            throw new ForbiddenException();
+        }
         BeanUtils.copyProperties(dto, event);
         eventRepository.save(event);
     }
 
     //Delete
-    public void deleteEvent(DeleteEventDTO dto) throws EventNotFoundException {
+    public void deleteEvent(DeleteEventDTO dto) throws EventNotFoundException, UserNotFoundException, ForbiddenException {
         var event = eventRepository.findById(dto.getUuid()).orElseThrow(EventNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(event.getId())){
+            throw new ForbiddenException();
+        }
         eventRepository.delete(event);
     }
